# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "L5 OpenApi",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "darius@matulionis.lt"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://apiv2.test/api",
            "description": "L5 Swagger OpenApi Server"
        }
    ],
    "paths": {
        "/article": {
            "get": {
                "tags": [
                    "Article"
                ],
                "summary": "Get list of articles",
                "description": "Returns list of all articles and its details",
                "operationId": "getArticleList",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Unsuccess"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Article"
                ],
                "summary": "Creates a new article",
                "description": "Creates a new article",
                "operationId": "storeArticle",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description",
                                    "image"
                                ],
                                "properties": {
                                    "title": {
                                        "description": "Article's title",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "Article's description",
                                        "type": "string"
                                    },
                                    "image": {
                                        "description": "Article's image",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success! Article created."
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        },
        "/article/{id}": {
            "get": {
                "tags": [
                    "Article"
                ],
                "summary": "Returns an article giving its id",
                "description": "Returns an article giving its id",
                "operationId": "showArticle",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Article id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operação com sucesso"
                    },
                    "400": {
                        "description": "Operação sem sucesso"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Article"
                ],
                "summary": "Updates an article",
                "description": "Updates an article",
                "operationId": "updateArticle",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Article id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description",
                                    "image",
                                    "user_id"
                                ],
                                "properties": {
                                    "title": {
                                        "description": "Article's title",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "Article's description",
                                        "type": "string"
                                    },
                                    "image": {
                                        "description": "Article's image",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "user_id": {
                                        "description": "Article's user",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Article"
                ],
                "summary": "Deletes an article giving its id",
                "description": "Deletes an article giving its id",
                "operationId": "deleteArticle",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Article id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operação com sucesso"
                    },
                    "400": {
                        "description": "Operação sem sucesso"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        },
        "/userarticle": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get list of articles with users info",
                "description": "Returns list of all articles, its details and its user info",
                "operationId": "getUsersandArticlesList",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Unsuccess"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "Password Based": {
                "type": "oauth2",
                "description": "Use a global client_id / client_secret and your username / password combo to obtain a token",
                "in": "header",
                "flows": {
                    "password": {
                        "authorizationUrl": "/oauth/authorize",
                        "tokenUrl": "/oauth/token",
                        "refreshUrl": "/oauth/token/refresh",
                        "scopes": []
                    }
                },
                "scheme": "https"
            }
        }
    }
}